Renamed favoritsm to favoritism.
