+++ jokes/favoritism.py
